Now strace can print PID, PPID, TGID like kdump
Works with -f(f) -o keys. In linux standard strace prints only TGID and
treats it as if it was the only identifier of a process. This patch
makes easier to debug multithreaded, forked programs (like apache's
httpd).
It is rather naive, so you're free to hack it on your own.